Added System::invalidateOutputs() virtual member. It is executed in place of System::operate() if System::inputsValid() returns false. The default behavior is to call Output<>::Value::setValueUndefined() on all of the Output<>::Value objects owned by that System.
This change required introducing an association between a System and a set of Output<>::Values. This is implemented in the same style as the relationship between Systems and Input<>s. There is an additional Output<> constructor parameter that is a pointer to the parent System, an abstract AbstractOutput::AbstractValue nested class to hold the reference to the System, and System keeps a list of the active Output<>::Values it owns.