Changed the way Inputs and Outputs handle information flow.
Before, info was pushed down the tree of Systems as it became available with a call to Output<>::Value::setValue().

Now a request for information is propagated up the tree of Systems with a call to Input<>::getValue() or Input<>::valueDefined(). System::operate() is called as needed to produce the requested data. This allows less bookkeeping and helps prevent situations in which calculations are performed in some System::operate() to produce data that is never used.

This reflects the view that a main-loop's goal is to produce outputs, not to consume inputs. Sometimes inputs need to be consumed to produce outputs, but the motivator is a request for information from a sink.