Added 'bool realtime = false' parameter to CommunicationsBus::receive() to distinguish between the way "normal" communication happens on a CommunicationsBus and the way a Wam uses a CommunicationsBus. The default implementation ignores this parameter.
Filled out stubs for the BusManager::receive() mechanism. BusManager::receive() is thread-safe and has the following behavior: If blocking and realtime are true, receive() will not return until it has received the message it expects. If blocking is true but realtime is false (the default), the function will periodically yield control to other threads that may also be waiting to receive messages. This behavior is largely untested.

Modified BusManager::enumerate() to use *this instead of this->bus as the CommunicationsBus.

Modified Puck::receiveGetPropertyReply() to use bus.receive() instead of bus.receiveRaw().

Added Puck::MIN_ID constant. Fixed small bug in BusManager::enumerate() that was trying to enumerate puck 0.