Added locking code to System, System::Input, and System::Output.
Added mandatoryCleanUp() methods to System, AbstractInput, AbstractOutput, Input, and Output. They are designed to be called from the most derived dtor possible. Their function is to cut the object in question out of the rest of the system in order to prevent threading issues such as calling pure virtual methods.

Classes that derive from System are now required to call System::mandatoryCleanUp() from their dtor.